Remove PHP response format
