+++ core/DataTable/Renderer.php
@@ -11,6 +11,7 @@ namespace Piwik\DataTable;
 use Exception;
 use Piwik\Columns\Dimension;
 use Piwik\Common;
+use Piwik\DataTable;
 use Piwik\Metrics;
 use Piwik\Piwik;
 use Piwik\BaseFactory;
@@ -143,8 +144,7 @@ abstract class Renderer extends BaseFactory
                                                  'json',
                                                  'csv',
                                                  'tsv',
-                                                 'html',
-                                                 'php'
+                                                 'html'
     );
 
     /**
@@ -378,5 +378,154 @@ abstract class Renderer extends BaseFactory
         }
 
         return $wrap;
+    }
+
+    /**
+     * Produces a flat php array from the DataTable, putting "columns" and "metadata" on the same level.
+     *
+     * For example, when  a originalRender() would be
+     *     array( 'columns' => array( 'col1_name' => value1, 'col2_name' => value2 ),
+     *            'metadata' => array( 'metadata1_name' => value_metadata) )
+     *
+     * a flatRender() is
+     *     array( 'col1_name' => value1,
+     *            'col2_name' => value2,
+     *            'metadata1_name' => value_metadata )
+     *
+     * @param null|DataTable|DataTable\Map|Simple $dataTable
+     * @return array  Php array representing the 'flat' version of the datatable
+     */
+    protected function convertDataTableToArray($dataTable = null)
+    {
+        if (is_null($dataTable)) {
+            $dataTable = $this->table;
+        }
+
+        if (is_array($dataTable)) {
+            $flatArray = $dataTable;
+            if (self::shouldWrapArrayBeforeRendering($flatArray)) {
+                $flatArray = array($flatArray);
+            }
+        } elseif ($dataTable instanceof DataTable\Map) {
+            $flatArray = array();
+            foreach ($dataTable->getDataTables() as $keyName => $table) {
+                $flatArray[$keyName] = $this->convertDataTableToArray($table);
+            }
+        } elseif ($dataTable instanceof Simple) {
+            $flatArray = $this->convertSimpleTable($dataTable);
+
+            reset($flatArray);
+            $firstKey = key($flatArray);
+
+            // if we return only one numeric value then we print out the result in a simple <result> tag
+            // keep it simple!
+            if (count($flatArray) == 1
+                && $firstKey !== DataTable\Row::COMPARISONS_METADATA_NAME
+            ) {
+                $flatArray = current($flatArray);
+            }
+        } // A normal DataTable needs to be handled specifically
+        else {
+            $array = $this->convertTable($dataTable);
+            $flatArray = $this->flattenArray($array);
+        }
+
+        return $flatArray;
+    }
+
+    /**
+     * Converts the given data table to an array
+     *
+     * @param DataTable $table
+     * @return array
+     */
+    protected function convertTable($table)
+    {
+        $array = [];
+
+        foreach ($table->getRows() as $id => $row) {
+            $newRow = array(
+                'columns'        => $row->getColumns(),
+                'metadata'       => $row->getMetadata(),
+                'idsubdatatable' => $row->getIdSubDataTable(),
+            );
+
+            if ($id == DataTable::ID_SUMMARY_ROW) {
+                $newRow['issummaryrow'] = true;
+            }
+
+            if (isset($newRow['metadata'][DataTable\Row::COMPARISONS_METADATA_NAME])) {
+                $newRow['metadata'][DataTable\Row::COMPARISONS_METADATA_NAME] = $row->getComparisons();
+            }
+
+            $subTable = $row->getSubtable();
+            if ($this->isRenderSubtables()
+                && $subTable
+            ) {
+                $subTable = $this->convertTable($subTable);
+                $newRow['subtable'] = $subTable;
+                if ($this->hideIdSubDatatable === false
+                    && isset($newRow['metadata']['idsubdatatable_in_db'])
+                ) {
+                    $newRow['columns']['idsubdatatable'] = $newRow['metadata']['idsubdatatable_in_db'];
+                }
+                unset($newRow['metadata']['idsubdatatable_in_db']);
+            }
+            if ($this->hideIdSubDatatable !== false) {
+                unset($newRow['idsubdatatable']);
+            }
+
+            $array[] = $newRow;
+        }
+        return $array;
+    }
+
+    /**
+     * Converts the simple data table to an array
+     *
+     * @param Simple $table
+     * @return array
+     */
+    protected function convertSimpleTable($table)
+    {
+        $array = [];
+
+        $row = $table->getFirstRow();
+        if ($row === false) {
+            return $array;
+        }
+        foreach ($row->getColumns() as $columnName => $columnValue) {
+            $array[$columnName] = $columnValue;
+        }
+
+        $comparisons = $row->getComparisons();
+        if (!empty($comparisons)) {
+            $array[DataTable\Row::COMPARISONS_METADATA_NAME] = $comparisons;
+        }
+
+        return $array;
+    }
+
+    /**
+     *
+     * @param array $array
+     * @return array
+     */
+    protected function flattenArray($array)
+    {
+        $flatArray = [];
+        foreach ($array as $row) {
+            $newRow = $row['columns'] + $row['metadata'];
+            if (isset($row['idsubdatatable'])
+                && $this->hideIdSubDatatable === false
+            ) {
+                $newRow += array('idsubdatatable' => $row['idsubdatatable']);
+            }
+            if (isset($row['subtable'])) {
+                $newRow += array('subtable' => $this->flattenArray($row['subtable']));
+            }
+            $flatArray[] = $newRow;
+        }
+        return $flatArray;
     }
 }
