+++ core/DataTable/Renderer/Xml.php
@@ -36,26 +36,6 @@ class Xml extends Renderer
     }
 
     /**
-     * Converts the given data table to an array
-     *
-     * @param DataTable|DataTable/Map $table  data table to convert
-     * @return array
-     */
-    protected function getArrayFromDataTable($table)
-    {
-        if (is_array($table)) {
-            return $table;
-        }
-
-        $renderer = new Php();
-        $renderer->setRenderSubTables($this->isRenderSubtables());
-        $renderer->setSerialize(false);
-        $renderer->setTable($table);
-        $renderer->setHideIdSubDatableFromResponse($this->hideIdSubDatatable);
-        return $renderer->flatRender();
-    }
-
-    /**
      * Computes the output for the given data table
      *
      * @param DataTable|DataTable/Map $table
@@ -66,7 +46,7 @@ class Xml extends Renderer
      */
     protected function renderTable($table, $returnOnlyDataTableXml = false, $prefixLines = '')
     {
-        $array = $this->getArrayFromDataTable($table);
+        $array = $this->convertDataTableToArray($table);
         if ($table instanceof Map) {
             $out = $this->renderDataTableMap($table, $array, $prefixLines);
 
@@ -188,11 +168,11 @@ class Xml extends Renderer
                 } else {
                     $result .= $prefixLines . $prefix . "\n";
                     if ($value instanceof Map) {
-                        $result .= $this->renderDataTableMap($value, $this->getArrayFromDataTable($value), $prefixLines);
+                        $result .= $this->renderDataTableMap($value, $this->convertDataTableToArray($value), $prefixLines);
                     } elseif ($value instanceof Simple) {
-                        $result .= $this->renderDataTableSimple($this->getArrayFromDataTable($value), $prefixLines);
+                        $result .= $this->renderDataTableSimple($this->convertDataTableToArray($value), $prefixLines);
                     } else {
-                        $result .= $this->renderDataTable($this->getArrayFromDataTable($value), $prefixLines);
+                        $result .= $this->renderDataTable($this->convertDataTableToArray($value), $prefixLines);
                     }
                     $result .= $prefixLines . $suffix . "\n";
                 }
@@ -381,7 +361,7 @@ class Xml extends Renderer
                 foreach ($row as $name => $value) {
                     // handle the recursive dataTable case by XML outputting the recursive table
                     if ($value instanceof DataTable) {
-                        $value = $this->getArrayFromDataTable($value);
+                        $value = $this->convertDataTableToArray($value);
                         if ($value instanceof Simple) {
                             $value = "\n" . $this->renderDataTableSimple($value, $prefixLine . "\t\t");
                         } else {
@@ -436,7 +416,7 @@ class Xml extends Renderer
             if (is_string($xmlValue) && strlen($xmlValue) == 0) {
                 $out .= $prefixLine . "\t<$tagStart />\n";
             } else if ($value instanceof DataTable || is_array($value)) {
-                $arrayValue = $this->getArrayFromDataTable($value);
+                $arrayValue = $this->convertDataTableToArray($value);
                 if (!is_array(reset($arrayValue))) {
                     $xmlTable = $this->renderDataTableSimple($arrayValue, $prefixLine . "\t");
                 } else {
