+++ plugins/API/API.php
@@ -524,7 +524,7 @@ class API extends \Piwik\Plugin\API
 
         $result = array();
         foreach ($urls as $url) {
-            $params = Request::getRequestArrayFromString($url . '&format=php&serialize=0');
+            $params = Request::getRequestArrayFromString($url . '&format=json');
 
             if (!empty($params['method']) && $params['method'] === 'API.getBulkRequest') {
                 continue;
@@ -543,7 +543,7 @@ class API extends \Piwik\Plugin\API
             }
 
             $req = new Request($params);
-            $result[] = $req->process();
+            $result[] = json_decode($req->process(), true);
         }
         return $result;
     }
