+++ plugins/SEO/Widgets/GetRank.php
@@ -51,13 +51,12 @@ class GetRank extends \Piwik\Widget\Widget
 
         $dataTable = API::getInstance()->getRank($url);
         
-        /** @var \Piwik\DataTable\Renderer\Php $renderer */
-        $renderer = Renderer::factory('php');
-        $renderer->setSerialize(false);
+        /** @var \Piwik\DataTable\Renderer\Json $renderer */
+        $renderer = Renderer::factory('json');
 
         return $this->renderTemplate('getRank', array(
             'urlToRank' => Url::getHostFromUrl($url),
-            'ranks' => $renderer->render($dataTable)
+            'ranks' => json_decode($renderer->render($dataTable), true)
         ));
     }
