+++ plugins/SEO/tests/Integration/SEOTest.php
@@ -51,9 +51,9 @@ class SEOTest extends IntegrationTestCase
         } catch(Exception $e) {
             $this->markTestSkipped('A SEO http request failed, Skipping this test for now. Error was: '.$e->getMessage());
         }
-        $renderer = Renderer::factory('php');
-        $renderer->setSerialize(false);
-        $ranks = $renderer->render($dataTable);
+        $renderer = Renderer::factory('json');
+        $renderer->setTable($dataTable);
+        $ranks = json_decode($renderer->render(), true);
         foreach ($ranks as $rank) {
             if ($rank["id"] == "alexa") { // alexa is broken at the moment
                 continue;
