--- tests/PHPUnit/System/expected/test_apiGetReportMetadata_phpRenderer__API.getDefaultMetricTranslations.php
@@ -1,1 +1,0 @@
-a:1:{i:0;a:112:{s:9:"nb_visits";s:6:"Visits";s:16:"nb_uniq_visitors";s:15:"Unique visitors";s:10:"nb_actions";s:7:"Actions";s:8:"nb_users";s:5:"Users";s:20:"nb_actions_per_visit";s:17:"Actions per Visit";s:16:"avg_time_on_site";s:20:"Avg. Time on Website";s:11:"bounce_rate";s:11:"Bounce Rate";s:15:"conversion_rate";s:15:"Conversion Rate";s:5:"label";s:5:"Label";s:4:"date";s:4:"Date";s:16:"avg_time_on_page";s:17:"Avg. time on page";s:14:"sum_time_spent";s:41:"Total time spent by visitors (in seconds)";s:16:"sum_visit_length";s:41:"Total time spent by visitors (in seconds)";s:12:"bounce_count";s:7:"Bounces";s:22:"bounce_count_returning";s:33:"Bounce Count for Returning Visits";s:11:"max_actions";s:28:"Maximum actions in one visit";s:21:"max_actions_returning";s:38:"Maximum actions in one returning visit";s:29:"nb_visits_converted_returning";s:36:"Number of converted returning visits";s:26:"sum_visit_length_returning";s:51:"Total time spent by returning visitors (in seconds)";s:19:"nb_visits_converted";s:23:"Visits with Conversions";s:14:"nb_conversions";s:11:"Conversions";s:7:"revenue";s:7:"Revenue";s:7:"nb_hits";s:9:"Pageviews";s:15:"entry_nb_visits";s:9:"Entrances";s:22:"entry_nb_uniq_visitors";s:16:"Unique entrances";s:14:"exit_nb_visits";s:5:"Exits";s:21:"exit_nb_uniq_visitors";s:12:"Unique exits";s:18:"entry_bounce_count";s:7:"Bounces";s:17:"exit_bounce_count";s:7:"Bounces";s:9:"exit_rate";s:9:"Exit rate";s:26:"sum_daily_nb_uniq_visitors";s:27:"Unique visitors (daily sum)";s:18:"sum_daily_nb_users";s:17:"Users (daily sum)";s:32:"sum_daily_entry_nb_uniq_visitors";s:28:"Unique entrances (daily sum)";s:31:"sum_daily_exit_nb_uniq_visitors";s:24:"Unique exits (daily sum)";s:16:"entry_nb_actions";s:27:"Actions after entering here";s:22:"entry_sum_visit_length";s:61:"Total time spent by visitors (in seconds) after entering here";s:12:"nb_pageviews";s:9:"Pageviews";s:17:"nb_uniq_pageviews";s:16:"Unique Pageviews";s:12:"nb_downloads";s:9:"Downloads";s:17:"nb_uniq_downloads";s:16:"Unique Downloads";s:11:"nb_outlinks";s:8:"Outlinks";s:16:"nb_uniq_outlinks";s:15:"Unique Outlinks";s:11:"nb_searches";s:8:"Searches";s:11:"nb_keywords";s:15:"Unique Keywords";s:19:"avg_time_generation";s:20:"Avg. generation time";s:19:"nb_pages_per_search";s:20:"Search Results pages";s:24:"nb_hits_following_search";s:25:"Clicked in search results";s:16:"visits_evolution";s:16:"Visits Evolution";s:17:"actions_evolution";s:17:"Actions Evolution";s:19:"pageviews_evolution";s:19:"Pageviews Evolution";s:17:"revenue_evolution";s:17:"Revenue Evolution";s:24:"nb_conversions_evolution";s:21:"Conversions Evolution";s:16:"orders_evolution";s:26:"Ecommerce Orders Evolution";s:27:"ecommerce_revenue_evolution";s:25:"Product Revenue Evolution";s:35:"Referrers_visitorsFromSearchEngines";s:28:"Visitors from Search Engines";s:43:"Referrers_visitorsFromSearchEngines_percent";s:39:"Percent of Visitors from Search Engines";s:36:"Referrers_visitorsFromSocialNetworks";s:29:"Visitors from Social Networks";s:44:"Referrers_visitorsFromSocialNetworks_percent";s:40:"Percent of Visitors from Social Networks";s:33:"Referrers_visitorsFromDirectEntry";s:26:"Visitors from Direct Entry";s:41:"Referrers_visitorsFromDirectEntry_percent";s:37:"Percent of Visitors from Direct Entry";s:30:"Referrers_visitorsFromWebsites";s:22:"Visitors from Websites";s:38:"Referrers_visitorsFromWebsites_percent";s:33:"Percent of Visitors from Websites";s:31:"Referrers_visitorsFromCampaigns";s:23:"Visitors from Campaigns";s:39:"Referrers_visitorsFromCampaigns_percent";s:34:"Percent of Visitors from Campaigns";s:31:"Referrers_distinctSearchEngines";s:23:"Distinct search engines";s:32:"Referrers_distinctSocialNetworks";s:24:"Distinct social networks";s:26:"Referrers_distinctWebsites";s:17:"Distinct websites";s:26:"Referrers_distinctKeywords";s:17:"Distinct keywords";s:27:"Referrers_distinctCampaigns";s:18:"Distinct campaigns";s:6:"orders";s:16:"Ecommerce Orders";s:17:"ecommerce_revenue";s:15:"Product Revenue";s:17:"revenue_per_visit";s:17:"Revenue per Visit";s:8:"quantity";s:8:"Quantity";s:9:"avg_price";s:13:"Average Price";s:12:"avg_quantity";s:16:"Average Quantity";s:16:"revenue_subtotal";s:8:"Subtotal";s:11:"revenue_tax";s:3:"Tax";s:16:"revenue_shipping";s:8:"Shipping";s:16:"revenue_discount";s:8:"Discount";s:17:"avg_order_revenue";s:19:"Average Order Value";s:9:"nb_events";s:6:"Events";s:15:"sum_event_value";s:11:"Event value";s:15:"min_event_value";s:19:"Minimum Event value";s:15:"max_event_value";s:19:"Maximum Event value";s:15:"avg_event_value";s:13:"Average value";s:20:"nb_events_with_value";s:19:"Events with a value";s:19:"nb_visits_returning";s:16:"Returning Visits";s:20:"nb_actions_returning";s:27:"Actions by Returning Visits";s:26:"avg_time_on_site_returning";s:43:"Avg. Duration of a Returning Visit (in sec)";s:21:"bounce_rate_returning";s:32:"Bounce Rate for Returning Visits";s:30:"nb_actions_per_visit_returning";s:32:"Avg. Actions per Returning Visit";s:26:"nb_uniq_visitors_returning";s:25:"Unique returning visitors";s:18:"nb_users_returning";s:15:"Returning Users";s:13:"nb_visits_new";s:10:"New Visits";s:14:"nb_actions_new";s:21:"Actions by New Visits";s:20:"avg_time_on_site_new";s:37:"Avg. Duration of a New Visit (in sec)";s:15:"bounce_rate_new";s:26:"Bounce Rate for New Visits";s:24:"nb_actions_per_visit_new";s:26:"Avg. Actions per New Visit";s:20:"nb_uniq_visitors_new";s:19:"Unique new visitors";s:12:"nb_users_new";s:9:"New Users";s:14:"nb_impressions";s:11:"Impressions";s:15:"nb_interactions";s:20:"Content Interactions";s:16:"interaction_rate";s:16:"Interaction Rate";s:20:"nb_visits_percentage";s:8:"% Visits";s:22:"nb_hits_with_bandwidth";s:28:"Recorded hits with bandwidth";s:13:"max_bandwidth";s:13:"Max Bandwidth";s:13:"min_bandwidth";s:13:"Min Bandwidth";s:13:"sum_bandwidth";s:13:"Sum Bandwidth";s:13:"avg_bandwidth";s:14:"Avg. Bandwidth";s:26:"nb_total_overall_bandwidth";s:25:"Bytes transferred overall";s:27:"nb_total_pageview_bandwidth";s:27:"Bytes transferred pageviews";s:27:"nb_total_download_bandwidth";s:27:"Bytes transferred downloads";}}
