--- tests/UI/expected-screenshots/ReportExporting_RowEvolution_exported.php.txt
@@ -1,1 +1,0 @@
-a:5:{i:2008;a:0:{}i:2009;a:0:{}i:2010;a:0:{}i:2011;a:0:{}i:2012;a:1:{i:0;a:14:{s:5:"label";s:12:"referrer.com";s:9:"nb_visits";i:11;s:10:"nb_actions";i:45;s:11:"max_actions";d:7;s:16:"sum_visit_length";i:7398;s:12:"bounce_count";i:2;s:19:"nb_visits_converted";i:11;s:5:"goals";a:2:{s:8:"idgoal=1";a:3:{s:14:"nb_conversions";i:23;s:19:"nb_visits_converted";i:11;s:7:"revenue";d:115;}s:8:"idgoal=2";a:3:{s:14:"nb_conversions";i:11;s:19:"nb_visits_converted";i:11;s:7:"revenue";d:55;}}s:14:"nb_conversions";i:34;s:7:"revenue";d:170;s:26:"sum_daily_nb_uniq_visitors";i:11;s:18:"sum_daily_nb_users";i:2;s:7:"segment";s:26:"referrerName==referrer.com";s:14:"idsubdatatable";i:8;}}}
